Integrate create change modal into manage changelog panel
