--- frontend/src/component/pages/shared/admin-tabs/ManageChangeLogs.scss
@@ -1,17 +1,0 @@
-@import '../../../styles/color';
-
-.manage-change-logs {
-  .create-change-button > .button {
-    background-color: $color-dark-gray;
-    padding: 10px 10px;
-    color: $color-light-gray;
-
-    :hover {
-      color: $color-white;
-    }
-  }
-
-  .modal .card {
-    width: 70%;
-  }
-}
